Move onJurisdictionClick handler to utils
